feat: utils updates for edge runtime compatability
