+++ packages/gasket-utils/lib/apply-config-overrides.js
@@ -1,25 +1,27 @@
-const defaultsDeep = require('lodash.defaultsdeep');
+// var defaultsDeep = require('lodash.defaultsdeep');
+const deepmerge = require('deepmerge');
 // @ts-ignore - diagnostics lib does not have a types declaration file
 const debug = require('diagnostics')('gasket:utils');
 
 /**
  * Normalize the config by applying any overrides for environments, commands,
  * or local-only config file.
- * @type {import('./index').applyConfigOverrides}
+ * @type {import('./apply-config-overrides').applyConfigOverrides}
  */
 function applyConfigOverrides(
   config,
   { env = '', commandId }
 ) {
-  return defaultsDeep(
-    {},
-    ...getPotentialConfigs(config, { env, commandId })
+  // @ts-ignore - merged config definitions
+  return deepmerge.all(
+  // @ts-ignore - partial config definitions
+    [...getPotentialConfigs(config, { env, commandId })].reverse()
   );
 }
 
 /**
  * Generator function to yield potential configurations
- * @type {import('./index').getPotentialConfigs}
+ * @type {import('./apply-config-overrides').getPotentialConfigs}
  */
 function *getPotentialConfigs(config, { env, commandId }) {
   // Separate environment-specific config from another config
@@ -34,7 +36,7 @@ function *getPotentialConfigs(config, { env, commandId }) {
 
 /**
  * Generator function to yield command overrides
- * @type {import('./internal').getCommandOverrides}
+ * @type {import('./apply-config-overrides').getCommandOverrides}
  */
 function *getCommandOverrides(commands, commandId) {
   const commandOverrides = commandId && commands[commandId];
@@ -46,7 +48,7 @@ function *getCommandOverrides(commands, commandId) {
 
 /**
  * Generator function to yield sub-environment overrides
- * @type {import('./internal').getSubEnvironmentOverrides}
+ * @type {import('./apply-config-overrides').getSubEnvironmentOverrides}
  */
 function *getSubEnvironmentOverrides(env, environments) {
   const envParts = env.split('.');
@@ -65,7 +67,7 @@ function *getSubEnvironmentOverrides(env, environments) {
 
 /**
  * Generator function to yield development overrides
- * @type {import('./internal').getDevOverrides}
+ * @type {import('./apply-config-overrides').getDevOverrides}
  */
 function *getDevOverrides(isLocalEnv, environments) {
   // Special case for the local environment, which inherits from the
@@ -79,4 +81,6 @@ function *getDevOverrides(isLocalEnv, environments) {
   }
 }
 
-module.exports = applyConfigOverrides;
+module.exports = {
+  applyConfigOverrides
+};
