+++ packages/gasket-utils/package.json
@@ -2,8 +2,6 @@
   "name": "@gasket/utils",
   "version": "7.0.6",
   "description": "Reusable utilities for Gasket internals",
-  "main": "lib",
-  "types": "lib/index.d.ts",
   "files": [
     "lib"
   ],
@@ -16,6 +14,17 @@
     "posttest": "npm run lint && npm run typecheck",
     "typecheck": "tsc",
     "typecheck:watch": "tsc --watch"
+  },
+  "exports": {
+    ".": {
+      "types": "./lib/index.d.ts",
+      "default": "./lib/index.js"
+    },
+    "./apply-config-overrides": {
+      "types": "./lib/apply-config-overrides.d.ts",
+      "default": "./lib/apply-config-overrides.js"
+    },
+    "./package.json": "./package.json"
   },
   "repository": {
     "type": "git",
@@ -37,16 +46,14 @@
   "dependencies": {
     "chalk": "^4.1.2",
     "concat-stream": "^2.0.0",
+    "deepmerge": "^4.3.1",
     "diagnostics": "^2.0.2",
-    "lodash.defaultsdeep": "^4.6.1",
     "semver": "^7.5.4"
   },
   "devDependencies": {
     "@gasket/core": "^7.0.6",
-    "@godaddy/dmd": "^1.0.4",
     "@types/concat-stream": "^2.0.3",
     "@types/cross-spawn": "^6.0.6",
-    "@types/lodash.defaultsdeep": "^4.6.9",
     "abort-controller": "^3.0.0",
     "cross-env": "^7.0.3",
     "eslint": "^8.56.0",
